Extract tx annotations from the enrichment service
... transactions.ts and utils.ts should maybe be rearranged, but we just
need a little more space right now.
